plugins/printcom/requirements.txt
ace_tools==0.0
ace_tools_open==0.1.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
dnspython==2.7.0
executing==2.2.0
fastapi[standard]==0.115.8
fonttools==4.55.8
h11==0.14.0
idna==3.10
ipython==8.32.0
itables==2.2.4
jedi==0.19.2
kiwisolver==1.4.8
matplotlib==3.10.0
matplotlib-inline==0.1.7
narwhals==1.25.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
plotly==6.0.0
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.11
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.45.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.29.0
wcwidth==0.2.13
awesome-slugify
